5758 - Add nationalDataPointLinks & split link types
