+++ src/meta/cycleData/links/descriptionLink.ts
@@ -1,0 +1,15 @@
+import { CommentableDescriptionName } from 'meta/assessment/descriptionValue'
+import { LinkLocationBase } from 'meta/cycleData/links/linkLocationBase'
+
+export const DescriptionLinkLocationPath = {
+  dataSourceReference: ['dataSources', 'reference'],
+  text: ['text'],
+}
+
+export type DescriptionsLocation = LinkLocationBase & {
+  colName: string
+  descriptionName: CommentableDescriptionName
+  path: Array<string>
+  sectionName: string
+  uuid?: string
+}
