+++ src/meta/cycleData/links/isDescriptionLocation.ts
@@ -1,4 +1,5 @@
-import { DescriptionsLocation, LinkLocation } from 'meta/cycleData/links/link'
+import { DescriptionsLocation } from 'meta/cycleData/links/descriptionLink'
+import { LinkLocation } from 'meta/cycleData/links/link'
 
 export const isDescriptionLocation = (location: LinkLocation): location is DescriptionsLocation => {
   return 'descriptionName' in location
