+++ src/meta/cycleData/links/linkLocationBase.ts
@@ -1,0 +1,4 @@
+export type LinkLocationBase = {
+  identifier?: string
+  url: string
+}
