+++ src/server/worker/tasks/verifyLinks/visitNationalDataPointLinks/props.ts
@@ -3,6 +3,7 @@ import { Job } from 'bullmq'
 import { CountryIso } from 'meta/area/countryIso'
 import { Assessment } from 'meta/assessment/assessment'
 import { Cycle } from 'meta/assessment/cycle'
+import { NDPLinkTarget } from 'meta/cycleData/links/nationalDataPointLink'
 
 import { VerifyLinksJobName } from 'server/worker/tasks/verifyLinks/jobNames'
 
@@ -11,8 +12,7 @@ export type VerifyNationalDataPointLinksJobProps = {
   countryIso: CountryIso
   cycle: Cycle
   notifyClients?: boolean
-  // TODO: targets: Array<NDPLinkTarget>
-  targets: Array<string>
+  targets: Array<NDPLinkTarget>
 }
 
 export type VerifyNationalDataPointLinksJob = Job<
